requirements.txt
certifi==2023.7.22
elastic-transport==8.4.1
elasticsearch==8.10.0
mysql-connector-python==8.1.0
protobuf==4.21.12
urllib3==1.26.17
